Fix segfault on open_file, change return types to bool
Open file could potentially segfault if the file pointer was null, aditionally the return types were confusing on both open_file and tc_load_config, to avoid problems all the return types are bool now, false defaults to error and true to success. Changed example to reflect on the return types and better handle errors, previously the error checking on example lead to a segfault as it was not safeguarding and not returning early, avoiding access to empty configs.