Update realloc strategy for tc_set_value
Prior to this update, while using tc_set_value, if the line size was bigger than the default line size we did a little magic to get a reallocation value, now a helper function gets the value location in the config and we can use strlen() on the the entire line and determine the exact size of reallocation if the new value provided overflows the default line size. Some empty spaces were removed from the file too and the copyright year updated.