Fix tc_save_to_file not saving anything to empty files
tc_save_to_file skipped any empty file when trying to write the configuration into it, thats because the open_file util function returned an error checking for an EOF when using fgetc(), in this case this is a bug because you could want to write the config file back to a newly created, empty file. For non-empty files the behavior was normal.